Add files via upload
User enters their username once and press 'get started'. This will create the user on the database so that the database can keep track of the jokes the user likes. Once the user enters his/her username, they don't have to do it again. I used UserDefaults to remember the user.